requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
camel-converter==4.0.1
cbor2==5.6.5
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.0
clerk-backend-api==1.6.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.3
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-permissions-policy==4.23.0
django-ratelimit==4.1.0
django-redis==5.4.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eval_type_backport==0.2.2
firebase-admin==6.6.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jsonpath-python==1.0.6
kombu==5.4.2
lxml==5.3.0
meilisearch==0.33.0
msgpack==1.1.0
mypy-extensions==1.0.0
oscrypto==1.3.0
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webauthn==2.4.0
webencodings==0.5.1
whitenoise==6.8.2
xhtml2pdf==0.2.16
